Added comments.

+++ AprioriClose/main.cpp
@@ -4,10 +4,14 @@
 
 int main()
 {
+    //Load the airlines csv
     ItemsetLoader itemset{"airlines.csv"};
     
-    itemset.Discretize<double>("Flight",10);
-    itemset.Discretize<double>("Time",10);
-    itemset.Discretize<double>("Length",10);
+    //Apply discretization on the required columns
+    itemset.Discretize<float>("Flight",10);
+    itemset.Discretize<float>("Time",10);
+    itemset.Discretize<float>("Length",10);
+
+    //Save the new document to disk
     itemset.SaveDocument("Airlines-discretized.csv");
 }
